requirements.txt
alembic==1.7.7
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
Flask==2.1.0
Flask-Humanize==0.3.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-paginate==2022.1.8
Flask-Pagination==0.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.11.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
psycopg2==2.9.3
pygeocoder==1.2.5
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
timeago==1.0.15
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
